Build live profile editing foundations
Freeze the Profile Wire v1 contracts and cross-language codecs, add dual-slot persistent storage and read-only firmware discovery, and integrate the Profile Studio HID transport and compatibility UI. Harden VIA split mirroring and replace misleading RP2040 resource assumptions with bank-aware accounting and tests.

+++ users/noah/lib/compat/qmk_via_split_mirror.c
@@ -46,7 +46,7 @@ static void noah_qmk_via_split_mirror_apply_command(const uint8_t *data, uint8_t
             }
 
             size = data[3];
-            if (length < (uint8_t)(4u + size)) {
+            if ((uint16_t)length < 4u + (uint16_t)size) {
                 return;
             }
 
@@ -61,7 +61,7 @@ static void noah_qmk_via_split_mirror_apply_command(const uint8_t *data, uint8_t
             }
 
             size = data[3];
-            if (length < (uint8_t)(4u + size)) {
+            if ((uint16_t)length < 4u + (uint16_t)size) {
                 return;
             }
